A MATLAB Project
Note 1: The noisy signals may be different in your plot from my report. It's all because of "rand()" function that we use to create noises. Don't worry guys. They are work correctly.
Note 2: Pay attention to this notice that the plots you get are the spectrum of their frequency. you can make an inverse of Fourier transform to make them in time. You could find some helps on MATLAB help. 
Good Luck